+++ modules/tecor/testframework/docs/demos/icst2018/build.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-echo "Building tecor demo"
-docker build -t vassik/tecor-demo:latest .
-echo "Pushing docker image to hub"
-docker push vassik/tecor-demo:latest